requirements.txt
pandas
pillow
pyzbar
pyotp
selenium
urllib3
protobuf
kiteconnect
click
kiteconnect
openpyxl
